refactored ProjectService
