+++ src/main/java/faang/school/projectservice/filter/InternshipFilter.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.service.filter;
+package faang.school.projectservice.filter;
 
 import faang.school.projectservice.dto.InternshipFilterDto;
 import faang.school.projectservice.model.Internship;
