+++ src/main/java/faang/school/projectservice/service/InternshipService.java
@@ -19,8 +19,8 @@ import faang.school.projectservice.repository.CandidateRepository;
 import faang.school.projectservice.repository.InternshipRepository;
 import faang.school.projectservice.repository.ProjectRepository;
 import faang.school.projectservice.repository.TeamMemberRepository;
-import faang.school.projectservice.service.filter.InternshipFilter;
-import faang.school.projectservice.validation.InternshipValidator;
+import faang.school.projectservice.filter.InternshipFilter;
+import faang.school.projectservice.validator.InternshipValidator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
