+++ src/main/java/faang/school/projectservice/service/VacancyService.java
@@ -11,8 +11,7 @@ import faang.school.projectservice.model.Candidate;
 import faang.school.projectservice.model.Team;
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.Vacancy;
-import faang.school.projectservice.repository.CandidateRepository;
-import faang.school.projectservice.service.filter.VacancyFilter;
+import faang.school.projectservice.filter.VacancyFilter;
 import faang.school.projectservice.validator.VacancyValidator;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
